some minor change to citizen management

+++ Xaysetha System/resident_add.cs
@@ -132,7 +132,7 @@ namespace Xaysetha_System
                 cmd.Parameters.AddWithValue("@religion", txtReligious.Text);
                 cmd.Parameters.AddWithValue("@dad_name", txtDadName.Text);
                 cmd.Parameters.AddWithValue("@mom_name", txtAddr.Text);
-                cmd.Parameters.AddWithValue("@family_book", int.Parse(txtFamBookNums.Text));
+                cmd.Parameters.AddWithValue("@family_book", BigInteger.Parse(txtFamBookNums.Text));
                 /*4th missing field*/cmd.Parameters.AddWithValue("@workplace", txtWorkplace.Text);
                 cmd.Parameters.AddWithValue("@citizenPics", memoryStream.ToArray());
                 cmd.Parameters.AddWithValue("@occupation", txtJobs.Text);
